+++ CHANGES.md
@@ -3,10 +3,11 @@ Changes and TODO
 
 ## 2012-????-?? / 0.2.0
 
+* [TODO] Reverse HTML entities
 * [TODO] ClojureScript support
 
 
-## 2012-July-?? / 0.1.0
+## 2012-August-?? / 0.1.0
 
 * Template with slots
 * Parse/compile and Render are separate
@@ -21,4 +22,6 @@ Changes and TODO
 * JVM specific API for template groups
 * Collection literals - vectors, sets, maps
 * Cache compiled templates (re-load only if modified)
-
+* HTML entities (escaping HTML to avoid XSS)
+* [TODO] Tests for JVM, HTML-entities and filter functions
+* [TODO] Documentation